Move function items out of pyobject module
