+++ vm/src/lib.rs
@@ -41,6 +41,7 @@ pub mod eval;
 mod exceptions;
 pub mod format;
 pub mod frame;
+pub mod function;
 pub mod import;
 pub mod obj;
 pub mod pyobject;
