+++ vm/src/stdlib/tokenize.rs
@@ -2,14 +2,14 @@
  * python tokenize module.
  */
 
-extern crate rustpython_parser;
 use std::iter::FromIterator;
 
-use self::rustpython_parser::lexer;
+use rustpython_parser::lexer;
 
+use crate::function::PyFuncArgs;
 use crate::obj::objstr;
-use crate::pyobject::{PyContext, PyFuncArgs, PyObjectRef, PyResult, TypeProtocol};
-use crate::VirtualMachine;
+use crate::pyobject::{PyContext, PyObjectRef, PyResult, TypeProtocol};
+use crate::vm::VirtualMachine;
 
 fn tokenize_tokenize(vm: &mut VirtualMachine, args: PyFuncArgs) -> PyResult {
     arg_check!(vm, args, required = [(readline, Some(vm.ctx.str_type()))]);
