+++ wasm/lib/src/browser_module.rs
@@ -1,16 +1,21 @@
-use crate::{convert, vm_class::AccessibleVM, wasm_builtins::window};
+use std::path::PathBuf;
+
 use futures::Future;
 use js_sys::Promise;
 use num_traits::cast::ToPrimitive;
+use wasm_bindgen::prelude::*;
+use wasm_bindgen::JsCast;
+use wasm_bindgen_futures::{future_to_promise, JsFuture};
+
+use rustpython_vm::function::PyFuncArgs;
+use rustpython_vm::import::import_module;
 use rustpython_vm::obj::{objint, objstr};
 use rustpython_vm::pyobject::{
-    AttributeProtocol, PyContext, PyFuncArgs, PyObject, PyObjectRef, PyResult, PyValue,
-    TypeProtocol,
+    AttributeProtocol, PyContext, PyObject, PyObjectRef, PyResult, PyValue, TypeProtocol,
 };
-use rustpython_vm::{import::import_module, VirtualMachine};
-use std::path::PathBuf;
-use wasm_bindgen::{prelude::*, JsCast};
-use wasm_bindgen_futures::{future_to_promise, JsFuture};
+use rustpython_vm::VirtualMachine;
+
+use crate::{convert, vm_class::AccessibleVM, wasm_builtins::window};
 
 enum FetchResponseFormat {
     Json,
